[JetBrains] Update Stable IDE images to 2022.3
