--- components/ide/jetbrains/backend-plugin/src/main/kotlin/io/gitpod/jetbrains/remote/GitpodCLIHelper.kt
@@ -1,11 +1,0 @@
-// Copyright (c) 2022 Gitpod GmbH. All rights reserved.
-// Licensed under the GNU Affero General Public License (AGPL).
-// See License.AGPL.txt in the project root for license information.
-
-package io.gitpod.jetbrains.remote
-
-import java.nio.file.Path
-
-interface GitpodCLIHelper {
-    suspend fun open(file: Path, shouldWait: Boolean)
-}
