+++ components/ide/jetbrains/backend-plugin/src/main/resources-latest/META-INF/extensions.xml
@@ -6,25 +6,5 @@
 <!--suppress PluginXmlValidity -->
 <idea-plugin>
     <extensions defaultExtensionNs="com.intellij">
-        <applicationService serviceInterface="io.gitpod.jetbrains.remote.GitpodIgnoredPortsForNotificationService" serviceImplementation="io.gitpod.jetbrains.remote.latest.GitpodIgnoredPortsForNotificationServiceImpl" preload="true"/>
-        <applicationService serviceInterface="io.gitpod.jetbrains.remote.GitpodCLIHelper" serviceImplementation="io.gitpod.jetbrains.remote.latest.GitpodCLIHelperImpl"/>
-        <applicationService serviceInterface="io.gitpod.jetbrains.remote.GitpodPortForwardingService" serviceImplementation="io.gitpod.jetbrains.remote.latest.GitpodPortForwardingServiceImpl" client="controller" preload="true"/>
-        <projectService serviceImplementation="io.gitpod.jetbrains.remote.GitpodClientProjectSessionTracker" client="controller" preload="true"/>
-        <projectService serviceImplementation="io.gitpod.jetbrains.remote.GitpodTerminalService" client="controller" preload="true"/>
     </extensions>
-    <actions>
-        <action id="io.gitpod.jetbrains.remote.latest.GitpodCopyUrlAction"
-                class="io.gitpod.jetbrains.remote.latest.GitpodCopyUrlAction"
-                text="Copy URL"
-                icon="AllIcons.Actions.Copy">
-            <add-to-group group-id="PortForwardingPortGroup" anchor="last"/>
-        </action>
-        <action id="io.gitpod.jetbrains.remote.latest.GitpodCopyWebUrlAction"
-                class="io.gitpod.jetbrains.remote.latest.GitpodCopyWebUrlAction"
-                text="Copy Web URL"
-                icon="AllIcons.Actions.Copy">
-            <add-to-group group-id="PortForwardingPortGroup" anchor="last"/>
-            <add-to-group group-id="PortForwardingSuggestionGroup" anchor="last"/>
-        </action>
-    </actions>
 </idea-plugin>
