+++ components/ide/jetbrains/backend-plugin/src/main/resources/META-INF/plugin.xml
@@ -26,15 +26,30 @@
         <applicationService serviceImplementation="io.gitpod.jetbrains.remote.services.HeartbeatService"
                             preload="true"/>
         <applicationService serviceImplementation="io.gitpod.jetbrains.remote.GitpodManager" preload="true"/>
+        <applicationService serviceInterface="io.gitpod.jetbrains.remote.GitpodIgnoredPortsForNotificationService"
+                            serviceImplementation="io.gitpod.jetbrains.remote.internal.GitpodIgnoredPortsForNotificationServiceImpl"
+                            preload="true"/>
+        <applicationService serviceInterface="io.gitpod.jetbrains.remote.GitpodPortForwardingService"
+                            serviceImplementation="io.gitpod.jetbrains.remote.internal.GitpodPortForwardingServiceImpl"
+                            client="controller" preload="true"/>
+
         <notificationGroup id="Gitpod Notifications" displayType="BALLOON" isLogByDefault="false"/>
+
         <httpRequestHandler implementation="io.gitpod.jetbrains.remote.GitpodCLIService"/>
+
         <projectService serviceImplementation="io.gitpod.jetbrains.remote.GitpodProjectManager" preload="true"/>
+        <projectService serviceImplementation="io.gitpod.jetbrains.remote.GitpodClientProjectSessionTracker"
+                        client="controller" preload="true"/>
+        <projectService serviceImplementation="io.gitpod.jetbrains.remote.GitpodTerminalService" client="controller"
+                        preload="true"/>
+
         <gateway.customization.name
                 implementation="io.gitpod.jetbrains.remote.GitpodGatewayClientCustomizationProvider"/>
         <gateway.customization.performance id="gitpodMetricsControl" order="before cpuControl"
                                            implementation="io.gitpod.jetbrains.remote.GitpodMetricControlProvider"/>
         <gateway.customization.metrics id="gitpodMetricsProvider"
                                        implementation="io.gitpod.jetbrains.remote.GitpodMetricProvider"/>
+
         <registryKey key="gitpod.autoJdk.disabled" defaultValue="false"
                      description="Disable auto-detection of JDK for the project and its modules"
                      restartRequired="true"/>
@@ -112,6 +127,19 @@
                 text="Gitpod: Follow Us On Twitter"
                 icon="AllIcons.Gutter.ExtAnnotation">
             <add-to-group group-id="UnattendedHostDropdownGroup" anchor="last"/>
+        </action>
+        <action id="io.gitpod.jetbrains.remote.actions.GitpodCopyUrlAction"
+                class="io.gitpod.jetbrains.remote.actions.GitpodCopyUrlAction"
+                text="Copy URL"
+                icon="AllIcons.Actions.Copy">
+            <add-to-group group-id="PortForwardingPortGroup" anchor="last"/>
+        </action>
+        <action id="io.gitpod.jetbrains.remote.actions.GitpodCopyWebUrlAction"
+                class="io.gitpod.jetbrains.remote.actions.GitpodCopyWebUrlAction"
+                text="Copy Web URL"
+                icon="AllIcons.Actions.Copy">
+            <add-to-group group-id="PortForwardingPortGroup" anchor="last"/>
+            <add-to-group group-id="PortForwardingSuggestionGroup" anchor="last"/>
         </action>
     </actions>
 </idea-plugin>
